Avoid `chmod()` when S3 Uploads is enabled
Gravity Forms uses `chmod` after renaming a temporary file to
ensure that permissions are set to 0644 by default. When S3 Uploads
is enabled, files are hosted on AWS S3 and `chmod()` is not a
supported stream operation in the AWS SDK that S3 Uploads uses.
In these cases, we're okay to skip `chmod()`.